research/models/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.3.0
tensorboard>=2.15.0
onnx>=1.15.0
onnxruntime>=1.16.0
coremltools>=7.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0
pandas>=2.1.0
